requirements.txt
blist>=1.3.6
cassandra-driver>=2.1.0b1
cliff>=1.6.1
jsonschema>=2.4.0
six>=1.7.3
statsd>=3.0.1
stevedore>=1.0.0.0a1
strict-rfc3339>=0.4
tornado>=4.0.2
hashids>=1.0.1
pycrypto>=2.6.1
python_jwt>=0.3.2
xmltodict>=0.9.0
